third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/iframe-javascript-url-more-encoding-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3CIFRAME%20src=%27javascript:alert%26%23x25%3B281)%27%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
